search-engine-main/requirements.txt
# ============================================================================
# CAMPUSHIRE RECRUITER-CANDIDATE MATCHING SYSTEM
# Complete Dependencies
# ============================================================================

# ============================================================================
# WEB FRAMEWORK
# ============================================================================
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
python-multipart>=0.0.10

# ============================================================================
# AI/ML AND EMBEDDINGS
# ============================================================================
# Gemini 2.0 SDK (Unified 2026 Version)
google-genai>=1.60.0
google-api-core>=2.20.0

# Vector operations and numerical computing
numpy>=1.24.0
scipy>=1.11.0

# ============================================================================
# DATABASE
# ============================================================================
# PostgreSQL adapter
psycopg2-binary>=2.9.0

# SQLAlchemy for ORM (optional, if needed later)
sqlalchemy>=2.0.0

# ============================================================================
# DATA VALIDATION & SERIALIZATION
# ============================================================================
pydantic>=2.10.0
pydantic-settings>=2.0.0

# ============================================================================
# SECURITY & AUTHENTICATION
# ============================================================================
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4

# ============================================================================
# ENVIRONMENT & CONFIGURATION
# ============================================================================
python-dotenv>=1.0.0

# ============================================================================
# LOGGING & MONITORING
# ============================================================================
# Built-in logging is sufficient, but these can enhance it
python-json-logger>=2.0.7

# ============================================================================
# PERFORMANCE & OPTIMIZATION
# ============================================================================
httpx[socks]>=0.28.0
aiofiles>=23.2.1

# ============================================================================
# DEVELOPMENT & TESTING DEPENDENCIES
# ============================================================================
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0

# ============================================================================
# CODE QUALITY
# ============================================================================
black>=23.9.0
flake8>=6.0.0
isort>=5.12.0
mypy>=1.5.0

# ============================================================================
# DOCUMENTATION
# ============================================================================
# FastAPI includes automatic docs generation via Swagger/ReDoc
